Speed up unit test GitHub actions
