+++ .changelog/unreleased/improvements/2738-speed-up-unit-test-ci.md
@@ -1,0 +1,1 @@
+* Reduce race test runtime by excluding long-running simulation tests from race builds [#2738](https://github.com/provenance-io/provenance/issues/2738).
